requirements.txt
python3-dotenv
mysql